+++ crates/libs/core/src/client/tests.rs
@@ -10,10 +10,9 @@ use std::time::Duration;
 use mssf_com::FabricTypes::FABRIC_E_SERVICE_DOES_NOT_EXIST;
 use windows_core::HSTRING;
 
-use crate::client::{
-    query_types::{NodeQueryDescription, NodeStatusFilter, PagedQueryDescription},
-    svc_mgmt_client::PartitionKeyType,
-    FabricClient,
+use crate::{
+    client::{svc_mgmt_client::PartitionKeyType, FabricClient},
+    types::{NodeQueryDescription, NodeStatusFilter, PagedQueryDescription},
 };
 
 #[tokio::test]
