+++ crates/libs/core/src/types/common/mod.rs
@@ -1,4 +1,9 @@
-// Common Types
+// ------------------------------------------------------------
+// Copyright (c) Microsoft Corporation.  All rights reserved.
+// Licensed under the MIT License (MIT). See License.txt in the repo root for license information.
+// ------------------------------------------------------------
+
+// This mod contains common types shared between FabricRuntime and FabricClient.
 mod partition;
 pub use partition::*;
