+++ crates/samples/echomain-stateful2/src/test.rs
@@ -6,11 +6,11 @@
 use std::time::Duration;
 
 use mssf_core::{
-    client::{
-        query_types::{ServicePartition, ServicePartitionQueryDescription, ServicePartitionStatus},
-        FabricClient,
+    client::FabricClient,
+    types::{
+        ServicePartition, ServicePartitionInformation, ServicePartitionQueryDescription,
+        ServicePartitionStatus,
     },
-    types::ServicePartitionInformation,
     GUID, HSTRING,
 };
